move conversion class into Utils file per Sambhav's suggestion
